+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## Unreleased
 
+## 2.5.0
+
+- feat: Update to latest Sentry SDKs (6.4.1) (#332)
+- fix: IPC breadcrumb serialization issue (#330)
+
+## 2.4.1
+
+- fix: Missing scope updates from isolated renderer (#322)
+- fix: Limit IPC serialization depth to 50 (#263)
+- fix: Check for require with typeof (#319)
+
 ## 2.4.0
 
 - feat: Add support for Electron 12